Create bg_trainer.py
Working prototype of a thing that creates a training dataset for scikit-learn. Currently: need to manually change which file you're saving to, on line 53 (e.g., 'dots.pkl' if you're clicking on dots, 'veins.pkl' if you're clicking on veins, 'noise.pkl' if you're clicking on featureless background). Also need to manually select the directory of the image.